+++ src/test/java/com/spectrum/service/PostServiceTest.java
@@ -9,7 +9,7 @@ import com.spectrum.controller.post.dto.PostListResponse;
 import com.spectrum.controller.post.dto.PostUpdateResponse;
 import com.spectrum.domain.post.Post;
 import com.spectrum.exception.post.PostNotFoundException;
-import com.spectrum.repository.PostRepository;
+import com.spectrum.repository.post.PostRepository;
 import com.spectrum.service.dto.PostCreateDto;
 import com.spectrum.service.dto.PostUpdateDto;
 import java.util.Optional;
